templates/agents/requirements.txt
anthropic>=0.42.0
httpx>=0.28.0
python-dotenv>=1.0.0
pyyaml>=6.0
tenacity>=8.2.3
